[AvatarGroup] Move from lab to core
