+++ packages/material-ui-lab/src/themeAugmentation/overrides.d.ts
@@ -1,4 +1,3 @@
-import { AvatarGroupClassKey } from '../AvatarGroup';
 import { TabListClassKey } from '../TabList';
 import { TabPanelClassKey } from '../TabPanel';
 import { TimelineClassKey } from '../Timeline';
@@ -8,13 +7,10 @@ import { TimelineDotClassKey } from '../TimelineDot';
 import { TimelineItemClassKey } from '../TimelineItem';
 import { TimelineOppositeContentClassKey } from '../TimelineOppositeContent';
 import { TimelineSeparatorClassKey } from '../TimelineSeparator';
-import { ToggleButtonClassKey } from '../ToggleButton';
-import { ToggleButtonGroupClassKey } from '../ToggleButtonGroup';
 import { TreeItemClassKey } from '../TreeItem';
 import { TreeViewClassKey } from '../TreeView';
 
 export interface LabComponentNameToClassKey {
-  MuiAvatarGroup: AvatarGroupClassKey;
   MuiTabList: TabListClassKey;
   MuiTabPanel: TabPanelClassKey;
   MuiTimeline: TimelineClassKey;
@@ -24,8 +20,6 @@ export interface LabComponentNameToClassKey {
   MuiTimelineItem: TimelineItemClassKey;
   MuiTimelineOppositeContent: TimelineOppositeContentClassKey;
   MuiTimelineSeparator: TimelineSeparatorClassKey;
-  MuiToggleButton: ToggleButtonClassKey;
-  MuiToggleButtonGroup: ToggleButtonGroupClassKey;
   MuiTreeItem: TreeItemClassKey;
   MuiTreeView: TreeViewClassKey;
 }
