--- packages/material-ui-lab/src/themeAugmentation/overrides.spec.ts
@@ -1,17 +1,0 @@
-import { createMuiTheme } from '@material-ui/core';
-
-createMuiTheme({
-  components: {
-    MuiAvatarGroup: {
-      styleOverrides: {
-        avatar: {
-          border: 'none',
-        },
-        // @ts-expect-error invalid class key
-        wrong: {
-          display: 'flex',
-        },
-      },
-    },
-  },
-});
