--- packages/material-ui-lab/src/themeAugmentation/props.spec.ts
@@ -1,13 +1,0 @@
-import { createMuiTheme } from '@material-ui/core';
-
-createMuiTheme({
-  components: {
-    MuiAvatarGroup: {
-      defaultProps: {
-        spacing: 'small',
-        // @ts-expect-error invalid prop
-        maximum: 5,
-      },
-    },
-  },
-});
