+++ packages/material-ui/src/AvatarGroup/AvatarGroup.test.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { expect } from 'chai';
 import { getClasses, createMount, describeConformance, createClientRender } from 'test/utils';
-import Avatar from '@material-ui/core/Avatar';
+import Avatar from '../Avatar';
 import AvatarGroup from './AvatarGroup';
 
 describe('<AvatarGroup />', () => {
