+++ packages/material-ui/src/AvatarGroup/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './AvatarGroup';
+export * from './AvatarGroup';
