+++ packages/material-ui/src/AvatarGroup/index.js
@@ -1,0 +1,1 @@
+export { default } from './AvatarGroup';
