+++ packages/material-ui/src/styles/props.d.ts
@@ -3,6 +3,7 @@ import { AlertTitleProps } from '../AlertTitle';
 import { AppBarProps } from '../AppBar';
 import { AutocompleteProps } from '../Autocomplete';
 import { AvatarProps } from '../Avatar';
+import { AvatarGroupProps } from '../AvatarGroup';
 import { BackdropProps } from '../Backdrop';
 import { BadgeProps } from '../Badge';
 import { BottomNavigationActionProps } from '../BottomNavigationAction';
@@ -123,7 +124,8 @@ export interface ComponentsPropsList {
   MuiAlertTitle: AlertTitleProps;
   MuiAppBar: AppBarProps;
   MuiAutocomplete: AutocompleteProps<any, any, any, any>;
-  MuiAvatar: AvatarProps;
+  MuiGroup: AvatarProps;
+  MuiAvatarGroup: AvatarGroupProps;
   MuiBackdrop: BackdropProps;
   MuiBadge: BadgeProps;
   MuiBottomNavigation: BottomNavigationProps;
